reviews

+++ reviews/models.py
@@ -1,8 +1,5 @@
 from django.db import models
 
-# class RatingField(models.DecimalField):
-    # decimal_places = 1
-    # max_digits = 2
 
 class Review(models.Model):
     owner = models.ForeignKey(
@@ -17,7 +14,7 @@ class Review(models.Model):
         default=None,
         related_name='reviews'
     )
-    rate = models.DecimalField(max_digits=3, decimal_places=1)
+    rate = models.IntegerField()
     date = models.DateTimeField(auto_now_add=True)
     text = models.TextField(max_length=300)
